src: Refactor post relation fetching
The shortcode method was a bit sluggish. This new one should get these
posts from the same JSON request for a post. This is temporary, as I'll
later be implementing a separate data type for subposts, since we can't
get post shortcodes anymore. Also, gets rid of the nested Option type
when fetching posts and users.