post: Implement Subpost
It's a lighter version of Post, a sort of "preview". Replaces the
Option<Vec<String>> type in the posts field for User, and in the parents
and replies fields for Post. Using the code field, you can fetch full
posts. Might implement conversion methods. Also, update test and docs
to reflect this change.