src: Unidirectional conversions
Add a conversion method for Author -> User, and gets rids of the
Post -> Subpost conversion, since they're unlikely to be used. Gets rid
of the fetch_post_code method too, since it's not needed for the
library. Remove the Default macro in all structs, since it's there's no
default for our purposes, and lastly update tests to reflect this.