meta: Bump version to v0.4
Let's sum up the changes in this new version. We simplified the output
of methods, implemented a dedicated error type using thiserror, added
Subpost and Author structs to embed within Post and User, implemented
methods within these new types to convert them into their detailed
counterparts, and cleaned up various parts of the code. This might
be one of the final iterations for a while.