post: Fetch replies within context
Previously, if the reply of a post was fetched, the replies field would
return all replies within the parent post. This adds a safety check,
where fetch_post compares the reply_to_author field of a subpost and,
if equal, saves it to the replies vector. Also, update the test to
check if replies are being fetched correctly after this check.