error: Deserialize errors from response
When an error occurs in a request, Threads returns a repsonse within an
`errors` array. Previously, this was not accounted for by fetch scripts,
instead assuming that these errors would mean the requested item was
just not found. This implements a new private function within the
SpoolsError enum, that takes the error's description, and responds with
a new element in the same enum: `ResponseError`.