meta: Bump version to v0.6.0
New spools release! It now fetches image thumbnails when parsing
requests, and improves on performance by taking advantage of Rust's
iterators and the fastrand crate for generating lsd IDs on requests.
Additionally, it now accounts for errors delivered directly by the
GraphQL endpoint, which adds a breaking change in the form of a new
ResponseError item in the SpoolsError enum.